requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0,<4.0
django-cors-headers
djangorestframework-simplejwt 
psycopg2-binary
python-decouple